# webhook/api/swagger-codegen/lettuce-webhook-openapi.yaml
openapi: 3.0.3
info:
  title: Plants Soil Lettuce (Webhook Engine)  - OpenAPI 3.0
  description: |-
    This is a Webhook Engine Core API based on the OpenAPI 3.0 specification.
    You can register as publisher and trigger (publish) events (message carried) as needed, all subscribers could receive the events instantly.
    You can register as subscriber with webhooks to listen on and receive events (messages carried) from publishers you subscribed.

    Some useful links:
    - [The Lettuce Webhook Engine repository](https://github.com/plants-soil/lettuce-webhook-hub)
  contact:
    email: danialdy@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: https://lettuce.plantssoil.com/api/v1
...
paths:
...
  /subscriber:
    post:
      tags:
        - subscriber
      summary: Add a new subscriber
      description: Add a new subscriber
      operationId: addSubscriber
      requestBody:
        description: Create a new subscriber
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Subscriber'
          application/xml:
            schema:
              $ref: '#/components/schemas/Subscriber'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Subscriber'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscriber'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Subscriber'
        '400':
          description: Invalid input
        '422':
          description: Validation exception
      security:
        - api_key: []
        - subscriber_auth:
            - write:subscribers
            - read:subscribers
  /subscriber/all:
    get:
      tags:
        - subscriber
      summary: Find all subscribers with pagination
      description: Will find the subscribers on the page specified (page, pageSize)
      operationId: findAllSubscribers
      parameters:
        - name: page
          in: query
          description: The page number of subscribers query result
          required: true
          explode: true
          schema:
            type: integer
            format: int32
        - name: pageSize
          in: query
          description: Maximum subscribers on current page
          required: true
          explode: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Subscriber'          
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Subscriber'
        '400':
          description: Validation exception
      security:
        - subscriber_auth:
            - write:subscribers
            - read:subscribers
  /subscriber/{subscriberId}:
    get:
      tags:
        - subscriber
      summary: Find subscriber by ID
      description: Returns a single subscriber
      operationId: findSubscriberById
      parameters:
        - name: subscriberId
          in: path
          description: ID of subscriber to return
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscriber'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Subscriber'
        '400':
          description: Invalid ID supplied
        '404':
          description: Subscriber not found
      security:
        - subscriber_auth:
            - write:subscribers
            - read:subscribers
    put:
      tags:
        - subscriber
      summary: Update an existing subscriber
      description: Update an existing subscriber by subscriberId
      operationId: updateSubscriberById
      parameters:
        - name: subscriberId
          in: path
          description: ID of subscriber that needs to be updated
          required: true
          schema:
            type: string
      requestBody:
        description: Update an existent subscriber
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Subscriber'
          application/xml:
            schema:
              $ref: '#/components/schemas/Subscriber'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Subscriber'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscriber'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Subscriber'
        '400':
          description: Required attributes cant't be null!
        '404':
          description: Subscriber not found
        '422':
          description: Validation exception
      security:
        - api_key: []
        - subscriber_auth:
            - write:subscribers
            - read:subscribers
    delete:
      tags:
        - subscriber
      summary: Deletes a subscriber
      description: delete a subscriber
      operationId: deleteSubscriber
      parameters:
        - name: api_key
          in: header
          description: ''
          required: false
          schema:
            type: string
        - name: subscriberId
          in: path
          description: ID of subscriber to delete
          required: true
          schema:
            type: string
      responses:
        '400':
          description: Invalid subscriber value
      security:
        - api_key: []
        - subscriber_auth:
            - write:subscribers
            - read:subscribers
...
components:
  schemas:
...
    Subscriber:
      description: Could be the subscriber which have multiple webhooks(applications) to subscribe publisher events, one organizaton only have one subscriber
      required:
        - organizationId
      type: object
      properties:
        subscriberId:
          type: string
          description: ID of subscriber, should be unique, system will assign auto-generated ID if it's null
          example: 6791d42ed3d172552ed09459, maximum 24 chars
        organizationId:
          type: string
          description: ID of organization which current subscriber should belongs to, required and should be unique
          example: 6791d42ed3d172552ed09459, maximum 24 chars
      xml:
        name: subscriber
...
  securitySchemes:
    organization_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://lettuce.plantssoil.com/oauth/authorize
          scopes:
            write:organizations: modify organizations in your account
            read:organizations: read your organizations
    publisher_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://lettuce.plantssoil.com/oauth/authorize
          scopes:
            write:publishers: modify publishers in you account
            read:publishers: read your publishers
    event_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://lettuce.plantssoil.com/oauth/authorize
          scopes:
            write:events: modify events in your account
            read:events: read your events
    datagroup_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://lettuce.plantssoil.com/oauth/authorize
          scopes:
            write:datagroups: modify data groups in your account
            read:datagroups: read your data groups
    subscriber_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://lettuce.plantssoil.com/oauth/authorize
          scopes:
            write:subscribers: modify subscribers in your account
            read:subscribers: read your subscribers
...